+++ openblas/cppbuild.sh
@@ -159,6 +159,13 @@ case $PLATFORM in
         export BINARY=64
         export TARGET=POWER5
         ;;
+    linux-mips64el)
+        export CC="$OLDCC -mabi=64"
+        export FC="$OLDFC -mabi=64"
+        export LDFLAGS="-Wl,-z,noexecstack"
+        export BINARY=64
+        export TARGET=MIPS
+        ;;
     linux-armhf)
         export CC="arm-linux-gnueabihf-gcc"
         export FC="arm-linux-gnueabihf-gfortran"